use list of float instead of double to avoid conversions
svn path=/branches/xmlplugin/; revision=206

+++ TKS-XML-plugin/XMLExtract/ChannelData.cs
@@ -10,6 +10,6 @@ namespace XmlExtract
     {
         public double Interval { get; set; }
         public double XOffset { get; set; }
-        public List<double> Data { get; set; }
+        public List<float> Data { get; set; }
     }
 }
